fix compile error on linux armv7.

+++ systemstat_linux.go
@@ -129,7 +129,7 @@ func getProcCPUSample() (s ProcCPUSample) {
 	syscall.Getrusage(syscall.RUSAGE_SELF, &processInfo)
 
 	s.Time = time.Now()
-	s.ProcMemUsedK = processInfo.Maxrss
+	s.ProcMemUsedK = int64(processInfo.Maxrss)
 	s.User = float64(processInfo.Utime.Usec)/1000000 + float64(processInfo.Utime.Sec)
 	s.System = float64(processInfo.Stime.Usec)/1000000 + float64(processInfo.Stime.Sec)
 	s.Total = s.User + s.System
